backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bidict==0.23.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
multidict==6.7.1
numpy==2.4.3
pandas==3.0.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
redis==7.4.0
scikit-learn==1.8.0
scipy==1.17.1
simple-websocket==1.1.0
six==1.17.0
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
wsproto==1.3.2
xgboost==3.2.0
yarl==1.23.0
